+++ csatar-plugins/csatar/lang/en/lang.php
@@ -183,6 +183,8 @@
                 'activeMandateDeleteError' => 'The Scout having the %name name has active Mandates, thus this Scout cannot be deleted.',
                 'scoutTeam' => 'Scout\'s team',
                 'inactivationWarning' => 'Please note that if you change status from active to inactive, scout\'s all mandates will expire!',
+                'bulkDeletePartialSuccess' => ':deletedCount deleted from :totalCount. ',
+                'bulkDeleteError' => 'The following errors occurred: ',
             ],
             'admin' => [
                 'menu' => [